1st quick update Feb 6th
Modified Auth guard to the following:

Stopped validating tokenIsValid as a void or null

Instead jwt.verify now verifies if there's no error, then userDecoded:any, will be assigned with the decode value. If there's an error, userDecoded will be undefined

Below, if userDecode is not undefined, validations will be made.

The purpose of this is the following:
When doing it as before, only sub was available, so if we wanted to validate using any other field, it would be impossible. This time, we are able to validate with username, email or the sub.