Changes Feb 6th - Lists
Created Service, Controller, Schema, Interface for List

List only contains a title, a board reference and a status.

Only members of the board where the list belongs to, are able to make changes or delete the list.

Created a method that allows the user to see all the lists that belong to a specific board based on its status (active or inaactive)